CriticalRate 자료형 int -> float 변환

+++ Source/HongtoChess/Champion/ChampionStatComponent.h
@@ -52,7 +52,7 @@ private:
 	int32 Range;
 
 	UPROPERTY(Transient, EditAnywhere, Category = "Stat", meta = (AllowPrivateAccess = true))
-	int32 CriticalRate;
+	float CriticalRate;
 
 	UPROPERTY(Transient, EditAnywhere, Category = "Stat", meta = (AllowPrivateAccess = true))
 	float CriticalCoefficient;
